requirements.txt
Django==5.2.2
djangorestframework
python-decouple
python-dotenv
requests
os